[SPARK-49130] Verify built images in `build-image` CI job via `docker run` test
This PR aims to verify built images in `build-image` CI job via `docker run` test

As we observed at #31 , the existing CI may build a corrupted image because we didn't verify it.
- #31

No, this is a test on our CI infra.

Pass the CIs. We can see the result of `docker run` like the following.
- https://github.com/apache/spark-kubernetes-operator/actions/runs/10274277967/job/28430624083?pr=32

![Screenshot 2024-08-06 at 14 38 11](https://github.com/user-attachments/assets/5eea4d44-335e-4733-b6b4-2fee3c139834)

No.

Closes #32 from dongjoon-hyun/SPARK-45923.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>