[SPARK-49319] Add `SparkCluster` to `spark-operator-api` module and examples
### What changes were proposed in this pull request?

This PR aims the following.
- Extend `spark-operator-api` module by adding `SparkCluster` in addition to `SparkApplication`.
- Provide target examples

Most `SparkCluster` API changes are a straight-forward extension from `SparkApplication`.

### Why are the changes needed?

To make `Apache Spark K8s Operator` support `Spark Cluster` on K8s environment in the same way.

### Does this PR introduce _any_ user-facing change?

No. This is a new feature which is not released yet.

### How was this patch tested?

This only added API layer.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #75 from dongjoon-hyun/SPARK-49319.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>