[SPARK-49348] Support `schedulerName` for SparkCluster
### What changes were proposed in this pull request?

This PR aims to support K8s Scheduler for SparkCluster via `spark.kubernetes.scheduler.name`.

### Why are the changes needed?

To support a submission to a custom scheduler like Apache YuniKorn.

### Does this PR introduce _any_ user-facing change?

No. This is a new feature.

### How was this patch tested?

Pass the CIs and manual testing. Without installing `YuniKorn` scheduler, the pod of SparkCluster will be in Pending status.
```
$ gradle build buildDockerImage spark-operator-api:relocateGeneratedCRD -x check
$ kubectl apply -f examples/cluster-on-yunikorn.yaml
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #89 from dongjoon-hyun/SPARK-49348.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>