[SPARK-49400] Use Gradle Version Catalog
### What changes were proposed in this pull request?

This PR aims to use a standard `Gradle Version Catalog`.
- https://docs.gradle.org/current/userguide/platforms.html#sub:conventional-dependencies-toml

### Why are the changes needed?

Currently, `gradle.properties` has all predefined versions, but it's not a good place.

Gradle provides a standard way, `Gradle Version Catalog`, via `libs.versions.toml`, which is supported by CIs and other tools. We had better follow the standard in order to simplify the code and reduce the maintenance burden.

```
- implementation("org.apache.spark:spark-kubernetes_$scalaVersion:$sparkVersion") {
+ implementation(libs.spark.kubernetes) {
```

### Does this PR introduce _any_ user-facing change?

No. This is a dev-only change.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #109 from dongjoon-hyun/SPARK-49400.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>