[SPARK-49623] Refactor prefix `appResources` to `workloadResources`
### What changes were proposed in this pull request?

This PR refactors previous `appResources` to `workloadResources` for helm chart / template / values to avoid confusion.

### Why are the changes needed?

In operator helm chart, prefix / group `appResources` was introduced to indicate the resources are created for Spark workloads whereas other resources are serving the operator deployment itself.

As we extend the support to cover `SparkCluster`, the naming becomes confusing as the resources in fact serve both `SparkApp` and `SparkCluster`. Thus, this PR aims to refactor the naming prefix to cover the new resource in a more general fashion.

### Does this PR introduce _any_ user-facing change?

No - not yet released

### How was this patch tested?

CI + Helm chart test.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #121 from jiangzho/helm.

Authored-by: zhou-jiang <zhou_jiang@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>