[SPARK-49706] Use `apache/spark` images instead of `spark`
### What changes were proposed in this pull request?

This PR aims to propose to use `apache/spark` images instead of `spark` because `apache/spark` images are published first. For example, the following are only available in `apache/spark` as of now.
- https://github.com/apache/spark-docker/pull/66
- https://github.com/apache/spark-docker/pull/67
- https://github.com/apache/spark-docker/pull/68

### Why are the changes needed?

To apply the latest bits earlier.

### Does this PR introduce _any_ user-facing change?

There is no change from `Apache Spark K8s Operator`.
Only the underlying images are changed.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #128 from dongjoon-hyun/SPARK-49706.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>