[SPARK-51507] Support creating config map that can be mounted by Spark pods for apps
### What changes were proposed in this pull request?

This PR introduces a new field `configMapSpecs`, which enables user to create configmap(s) which can be later mounted on driver and/or executors with values specified in SparkApplications. It also adds corresponding docs and unit tests.

### Why are the changes needed?

There's a common user demand for the feature to create lightweight values / files override using configmap. This new feature saves the trouble for user to create configmap(s) beforehand - the config map would share the lifecycle of the corresponding Spark app.

User may mount the ConfigMap as needed in driver and executor pod templates - one example is provided in the doc. In future we may further enhance this flow by mounting the created config map automatically to Spark pods. This can be achieved by supporting volume type of ConfigMap in Spark Kubernetes, or by adding mutating webhook & alter pod templates after creation. For now, this PR opens up the possibility of creating config map and mounting them with one single spec.

### Does this PR introduce _any_ user-facing change?

No - not yet released

### How was this patch tested?

Pass the CIs

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #166 from jiangzho/cm.

Authored-by: Zhou JIANG <jiangzho@umich.edu>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>