[SPARK-52118] Enforce `Helm Chart` configuration value verification
### What changes were proposed in this pull request?

This PR aims to elevate the quality and maintainability of `Helm Chart` by introducing a layer of strict validation for user-provided configurations.

### Why are the changes needed?

By providing `values.schema.json` file, we can define and validate the expected configuration values that a user can provide to customize a chart's deployment. This will help the users by preventing mistakes and improving reliability.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #196 from dongjoon-hyun/SPARK-52118.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>