[SPARK-52304] Add `pi-with-driver-timeout` example
### What changes were proposed in this pull request?

This PR aims to add `pi-with-driver-timeout` example.

### Why are the changes needed?

`spark.driver.timeout` with `DriverTimeoutPlugin` is a new feature of Apache Spark 4.0.0 in order to guarantee the maximum run time of the jobs in the best effort approach.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manual test the example (timeout: 1minutes)

```
$ kubectl apply -f examples/pi-with-driver-timeout.yaml

$ kubectl get sparkapp
NAME                     CURRENT STATE                       AGE
pi-with-driver-timeout   TerminatedWithoutReleaseResources   2m59s
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #228 from dongjoon-hyun/SPARK-52304.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>