[SPARK-52474] Limit GHA job execution time to up to 20 minutes
### What changes were proposed in this pull request?

This PR aims to limit GHA job execution time to up to 20 minutes.

### Why are the changes needed?

According to the GitHub Action log on `main` branch, most CI finishes in 10 minutes. So, we can use `20` minutes as a job execution limit. This will stop abnormal executions. In those cases, we had better restart quickly.
- https://github.com/apache/spark-kubernetes-operator/actions/workflows/build_and_test.yml?query=branch%3Amain

### Does this PR introduce _any_ user-facing change?

No, this is a infra-only change.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #244 from dongjoon-hyun/SPARK-52474.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>