[SPARK-52581] Revise AppCleanUpStep to include cleanup logic for all states
### What changes were proposed in this pull request?

This PR updates AppCleanUpStep so that it can be added as a regular step of each reconciliation, remove the therefore-unused AppTerminatedStep

### Why are the changes needed?

Previously, we have AppTerminatedStep performed at the beginning of each reconciliation, to end the actions early if app has reached terminated state.

However, in some scenarios, we may still want to do "clean up" for applications that are terminated:

* If an app in "terminated without releasing resources" is being deleted, we would like the AppCleanUpStep to be performed, since it covers a few corner cases that might not be covered in general delete flow
* We may also to introduce the concept of "maximal retention duration" for applications along with current "retention policy" in near future to serve the lifecycle management of application secondary resources. That could also require cleaning up secondary resources for already-terminated applications.

The revised AppCleanUpStep includes everything in the previous AppTerminatedStep and takes care of cleaning up resources for those in "TerminatedWithoutReleaseResources" stage as needed.

### Does this PR introduce any user-facing change?

No - internal step(s) changes only

### How was this patch tested?

CIs and E2Es shall cover the clean up logic

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #257 from jiangzho/cleanupForCancel.

Authored-by: Zhou JIANG <jiangzho@umich.edu>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>