[SPARK-52645] Enable `ExitOnOutOfMemoryError` JVM option by default
### What changes were proposed in this pull request?

This Pr aims to enable `ExitOnOutOfMemoryError` JVM option by default.

### Why are the changes needed?

To help users identify the problematic situation as earlier as possible.

### Does this PR introduce _any_ user-facing change?

Yes, but OOM should not be ignored.

After this PR, the JVM will exit on the first `OutOfMemoryError` it encounters, terminating the process with exit code 3.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #266 from dongjoon-hyun/SPARK-52645.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>