[SPARK-52662] Add `SparkApp` benchmark
### What changes were proposed in this pull request?

This PR aims to add a simple E2E benchmark for `SparkApp` as a baseline.

### Why are the changes needed?

To check `SparkApp` creation and deletion in a measurable way.

### Does this PR introduce _any_ user-facing change?

No, this is a test.

### How was this patch tested?

Manual tests. When the argument is not given, the default value for the number of jobs is 1000.

```
$ sh tests/benchmark/sparkapps.sh 100
CLEAN UP NAMESPACE FOR BENCHMARK
START BENCHMARK WITH 100 JOBS
FINISHED 100 JOBS IN 56 SECONDS.
DELETED 100 JOBS IN 40 SECONDS.
```

```
$ sh tests/benchmark/sparkapps.sh
CLEAN UP NAMESPACE FOR BENCHMARK
START BENCHMARK WITH 1000 JOBS
FINISHED 1000 JOBS IN 340 SECONDS.
DELETED 1000 JOBS IN 401 SECONDS.
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #273 from dongjoon-hyun/SPARK-52662.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>