[SPARK-53409] Upgrade `Gradle` to 9.0.0
### What changes were proposed in this pull request?

This PR aims to upgrade `Gradle` to 9.0.0.

### Why are the changes needed?

`Gradle 9` is a new major release based on Java 17. We had better start with `Gradle 9` at `Spark K8s Operator v1.0.0`.

- https://docs.gradle.org/9.0.0/release-notes.html

### Does this PR introduce _any_ user-facing change?

No, this is a dev-only change.

### How was this patch tested?

Pass the CIs and check the following log.

```
Welcome to Gradle 9.0.0!

Here are the highlights of this release:
 - Configuration Cache is the recommended execution mode
 - Gradle requires JVM 17 or higher to run
 - Build scripts use Kotlin 2.2 and Groovy 4.0
 - Improved Kotlin DSL script compilation avoidance
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #301 from dongjoon-hyun/SPARK-53409.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>