[SPARK-53510] Update `SparkClusterResourceSpec` to use `apache/spark:4.0.1` as the default value of container image
### What changes were proposed in this pull request?

This PR aims to update `SparkClusterResourceSpec` to use `apache/spark:4.0.1` as the default value of container image.

### Why are the changes needed?

Since Apache Spark 4.0.1 image is published, we had better use this bug fixed version.

### Does this PR introduce _any_ user-facing change?

Yes if the user didn't specify the container image for Spark Cluster, but this is an expected change in that case.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #311 from dongjoon-hyun/SPARK-53510.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>