[SPARK-53627] Update docs to recommend K8s 1.32+
### What changes were proposed in this pull request?

This PR aims to update documentations to recommend K8s v1.32+.

### Why are the changes needed?

K8s 1.31 entered maintenance mode already on August 28, 2025 and will reach the `End of Life` date on Oct 28, 2025. We had better recommend users to use K8s v1.32+.

- https://kubernetes.io/releases/patch-releases/#1-31

In addition, the default K8s versions of public cloud providers are already moving to K8s 1.32+ like the following.

- EKS: v1.32 (Default), v1.33 (Available)
- AKS: v1.32 (Default), v1.33 (Available), v1.34 (Preview, will be GA on October 2025)
- GKE: v1.33 (Stable), v1.34 (Regular on 9/30), v1.34 (Rapid)

### Does this PR introduce _any_ user-facing change?

No because this is a documentation-only change.

### How was this patch tested?

Manual review because the result of CIs is irrelevant.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #321 from dongjoon-hyun/SPARK-53627.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>