[SPARK-53642] Improve `Javadoc` for interfaces, classes, methods, variables
### What changes were proposed in this pull request?

This PR aims to improve `Javadoc` for interfaces, classes, methods, and variables as a part of v0.5 preparation.

### Why are the changes needed?

Currently, we have comments partially. We had better be consistent as much as possible.

- Fill many javadoc comments.
- Fill missing `param` descriptions.
- Fix typos.
- Unify the description style like a sentence, e.g., adding a period at the end.

After making this is a baseline, we can add and revise more in the future at every releases.

### Does this PR introduce _any_ user-facing change?

No behavior change because this is a documentation only change.

### How was this patch tested?

Pass the CIs (including building `javadoc` successfully).

### Was this patch authored or co-authored using generative AI tooling?

Yes, I asked `gemini-2.5-flash` to fill the comments first and revised.

Closes #324 from dongjoon-hyun/SPARK-53642.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>