[SPARK-53644] Upgrade `Gradle` to 9.1.0
### What changes were proposed in this pull request?

This PR aims to upgrade `Gradle` to 9.1.0.

### Why are the changes needed?

`Gradle 9.1` is a new feature release with `Full Java 25 Support`.

- https://docs.gradle.org/9.1.0/release-notes.html

### Does this PR introduce _any_ user-facing change?

No, this is a dev-only change.

### How was this patch tested?

Pass the CIs and check the following log.

```
Welcome to Gradle 9.1.0!
Here are the highlights of this release:
 - Full Java 25 support
 - Native task graph visualization
 - Enhanced console output
For more details see https://docs.gradle.org/9.1.0/release-notes.html
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #325 from dongjoon-hyun/SPARK-53644.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>